bug: store score as int32_t instead of int16_t
Given the search depth I started notifying that the "good moves"
starting being extremely bad. This could ofc mean overflow which was in
fact the case.
I was a bit naive with my previous implementation. Instead increase the
scoring to use 32bit value to avoid over/underflows.
Also printing the score from the evaluator always returning 0. Also
fixed that bug.

+++ src/board_defs.h
@@ -92,16 +92,16 @@ constexpr static inline uint8_t s_secondRow { 8 };
 constexpr static inline uint8_t s_seventhRow { 48 };
 constexpr static inline uint8_t s_eightRow { 56 };
 
-constexpr static inline uint16_t s_maxMoves { 256 };
+constexpr static inline uint32_t s_maxMoves { 256 };
 constexpr static inline uint64_t s_aFileMask { 0x0101010101010101 };
 constexpr static inline uint64_t s_hFileMask { 0x8080808080808080 };
 
 constexpr static inline uint64_t s_row2Mask { 0xffULL << s_secondRow };
 constexpr static inline uint64_t s_row7Mask { 0xffULL << s_seventhRow };
 
-constexpr static inline int16_t s_maxScore = std::numeric_limits<int16_t>::max();
-constexpr static inline int16_t s_minScore = -s_maxScore; // num limit is +1 higher than max
+constexpr static inline int32_t s_maxScore = std::numeric_limits<int32_t>::max();
+constexpr static inline int32_t s_minScore = -s_maxScore; // num limit is +1 higher than max
 
-constexpr static inline int16_t s_mateValue { s_maxScore - 1000 };
-constexpr static inline int16_t s_mateScore { s_maxScore - 2000 };
+constexpr static inline int32_t s_mateValue { s_maxScore - 1000 };
+constexpr static inline int32_t s_mateScore { s_maxScore - 2000 };
